requirements.txt
tweepy
pandas
nltk
vaderSentiment
spacy
folium
matplotlib
seaborn
python-dotenv